+++ app/(tabs)/(more)/index.tsx
@@ -5,57 +5,56 @@ import { router } from 'expo-router';
 
 import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
+import { useColors } from '@/hooks/useColors';
 
-const MoreScreen = () => {
+export default function MoreScreen() {
+  const { tintColor } = useColors();
   return (
     <ThemedView style={styles.container}>
       <Pressable
         onPress={() => router.push('/about')}
-        style={styles.linkContainer}
+        style={[styles.button, { shadowColor: tintColor }]}
       >
-        <ThemedText style={styles.link}>حول التطبيق</ThemedText>
+        <ThemedText type="title" style={styles.buttonText}>
+          حول التطبيق
+        </ThemedText>
       </Pressable>
-      <Pressable
-        onPress={() => router.push('/privacy')}
-        style={styles.linkContainer}
-      >
-        <ThemedText style={styles.link}>سياسة الخصوصية</ThemedText>
+      <Pressable onPress={() => router.push('/privacy')} style={styles.button}>
+        <ThemedText type="title" style={styles.buttonText}>
+          سياسة الخصوصية
+        </ThemedText>
       </Pressable>
-      <Pressable
-        onPress={() => router.push('/settings')}
-        style={styles.linkContainer}
-      >
-        <ThemedText style={styles.link}>الإعدادات</ThemedText>
+      <Pressable onPress={() => router.push('/settings')} style={styles.button}>
+        <ThemedText type="title" style={styles.buttonText}>
+          الإعدادات
+        </ThemedText>
       </Pressable>
     </ThemedView>
   );
-};
+}
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    display: 'flex',
+    gap: 20,
     justifyContent: 'center',
     alignItems: 'center',
-    paddingHorizontal: 20,
-    paddingVertical: 40,
   },
-  linkContainer: {
-    marginVertical: 12,
-    width: '100%',
-    paddingVertical: 15,
-
-    borderRadius: 10,
-    elevation: 2, // For subtle shadow
-    alignItems: 'center',
+  button: {
+    width: '90%',
+    borderRadius: 5,
+    backgroundColor: '#4A90E2',
+    shadowOffset: { width: 0, height: 5 },
+    shadowOpacity: 0.2,
+    shadowRadius: 5,
+    elevation: 5,
   },
-  link: {
-    fontSize: 22,
-    paddingVertical: 10,
-    color: '#1E90FF',
+  buttonText: {
+    color: '#FFFFFF',
     textAlign: 'center',
-    fontWeight: '500',
-    letterSpacing: 0.5,
+    fontSize: 18,
+    fontWeight: '600',
+    paddingVertical: 10,
   },
 });
-
-export default MoreScreen;